Fixed: No pointer given to remove list func, added test.cor to gitignore

+++ lib/linked_list/src/destroy_list.c
@@ -13,7 +13,8 @@ void destroy_list(list_t *list, void (*destroy_data)(void *data))
     list_node_t *temp = NULL;
 
     for (node = list->head; node; node = temp) {
-        destroy_node(node, destroy_data);
+        if (destroy_data)
+            destroy_node(node, destroy_data);
         temp = node->next;
         free(node);
     }
